requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
streamlit==1.32.0
pandas==2.0.3
watchdog==3.0.0
boto3==1.37.0
uuid==1.30 